images/s7comm/client/requirements.txt
python-snap7
dnspython
